+++ test/base-functor/Main.hs
@@ -6,12 +6,11 @@ import Covenant.Data
     hasRecursive,
     mkBaseFunctor,
   )
-import Covenant.Test
+import Covenant.Unsafe
   ( DataDeclFlavor (Poly1PolyThunks),
     DataDeclSet (DataDeclSet),
     prettyDeclSet,
   )
-import Covenant.Type ()
 import Data.Map.Strict qualified as M
 import Optics.Core (view)
 import Test.QuickCheck